feat(e2): simetria dos dois destinos do datamart (T2.4)
Base RepositoryDatamart com read/transform; RepositoryGoldDatamart e
RepositoryDatamartClickhouse herdam dela e definem apenas __init__ e write. Antes o braco
ClickHouse era copia colada do Postgres: duas copias que divergissem depois passariam numa
conferencia de olho, e a comparacao mediria a diferenca de codigo em vez da do motor. Agora
e assercao executavel — read e transform sao o MESMO objeto de funcao nos dois.

Nao e heranca de PipelineGold, como a spec v3.0 previa: na 3.3.0 o braco Postgres e um
Repository proprio e PipelineGold grava Delta. A base fica onde read/transform vivem.

A limpeza de nulo entra dirigida por colunas_obrigatorias, nao como constante de classe:
["timestamp","filial","banco","unidade_producao_id"] e especifico de fact_200_cep, e fixar
isso quebraria na_drop em toda tabela gold sem essas colunas. Vazio e o default e desliga a
limpeza, entao gold_datamart em producao nao muda. Chega por --colunas_obrigatorias, por
--tables_json ou pela seed do Mongo.

CAST(... AS DECIMAL(9,3)) nos 4 casts da query, inclusive o do WHERE, para o filtro operar
sobre o mesmo valor que a projecao. O get_type_sql traduz DecimalType para NUMERIC(9,3), e
o ClickHouse ja declara Decimal(9,3): um lugar so define o tipo nos dois destinos.

Chave datamart_pg como apelido de gold_datamart, para a DAG ficar simetrica no nome.

Aceite: 190 passed na unidade, 6 passed contra o Postgres real e 6 contra o ClickHouse real
— o sexto caso do ClickHouse cobre timestamp nulo descartado antes da troca de particao.

+++ spark-source-code/resources/queries/fact_200_cep.sql
@@ -17,11 +17,11 @@ SELECT
         'yyyy-MM-dd HH:mm:ss.SSS'
     ) AS timestamp,
     'Peso SET' AS atributo_nome_pai,
-    CAST(dap.real AS DOUBLE) AS valor,
+    CAST(dap.real AS DECIMAL(9,3)) AS valor,
     'Peso MAX' AS nome_limite_superior,
     'Peso MIN' AS nome_limite_inferior,
-    CAST(dap.limite_superior AS DOUBLE) AS valor_limite_superior,
-    CAST(dap.limite_inferior AS DOUBLE) AS valor_limite_inferior,
+    CAST(dap.limite_superior AS DECIMAL(9,3)) AS valor_limite_superior,
+    CAST(dap.limite_inferior AS DECIMAL(9,3)) AS valor_limite_inferior,
     '' AS unidade_medida,
     '' AS atributo_tipo
 FROM delta.`MINIO_BASE_PATH/S3_PATH_SILVER/dw_andon_peso` dap
@@ -46,4 +46,4 @@ WHERE
         substring(dap.data_hora, 1, 23),
         'yyyy-MM-dd HH:mm:ss.SSS'
     ) < to_timestamp('JANELA_FIM')
-    AND CAST(dap.real AS DOUBLE) <> 0
+    AND CAST(dap.real AS DECIMAL(9,3)) <> 0
